# kubernetes/swagger-config copy.yaml
openapi: 3.0.3
info:
  title: Messenger API
  version: 1.0.1
  description: API мессенджера
servers:
  - url: /api
    description: REST API сервера
  - url: ws://localhost:8083
    description: WebSocket сервер для звонков
security:
  - BearerAuth: []
...
paths:
...
  /v1/rooms:
    get:
      tags:
        - Комнаты
      summary: Получить список всех комнат пользователя
      description: Возвращает список комнат, в которых участвует пользователь
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список комнат
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
        '401':
          description: Отсутствует токен авторизации
        '500':
          description: Ошибка при получении списка комнат
    post:
      tags:
        - Комнаты
      summary: Создать новую комнату
      description: Создает новую комнату для звонков
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Название комнаты
                  example: "Моя комната"
      responses:
        '201':
          description: Комната успешно создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
                  data:
                    $ref: '#/components/schemas/Room'
        '401':
          description: Отсутствует токен авторизации
        '500':
          description: Ошибка при создании комнаты
  /v1/rooms/{roomId}:
    get:
      tags:
        - Комнаты
      summary: Получить информацию о комнате
      description: Возвращает детальную информацию о комнате
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID комнаты
      responses:
        '200':
          description: Информация о комнате
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/Room'
        '401':
          description: Отсутствует токен авторизации
        '403':
          description: Нет доступа к комнате
        '404':
          description: Комната не найдена
        '500':
          description: Ошибка при получении информации о комнате
...
  /v1/rooms/{roomId}/join:
    post:
      tags:
        - Комнаты
      summary: Присоединиться к комнате по ID
      description: Присоединяется к комнате, используя её ID
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID комнаты
      responses:
        '200':
          description: Успешное присоединение к комнате
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/Room'
        '401':
          description: Отсутствует токен авторизации
        '404':
          description: Комната не найдена
        '500':
          description: Ошибка при присоединении к комнате
  /v1/rooms/{roomId}/leave:
    post:
      tags:
        - Комнаты
      summary: Покинуть комнату
      description: Покидает указанную комнату
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID комнаты
      responses:
        '200':
          description: Успешный выход из комнаты
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
        '401':
          description: Отсутствует токен авторизации
        '404':
          description: Участник не найден в комнате
        '500':
          description: Ошибка при выходе из комнаты
  /v1/rooms/{roomId}/participants:
    get:
      tags:
        - Комнаты
      summary: Получить список участников комнаты
      description: Возвращает список всех участников комнаты
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID комнаты
      responses:
        '200':
          description: Список участников
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoomParticipant'
        '401':
          description: Отсутствует токен авторизации
        '403':
          description: Нет доступа к комнате
        '500':
          description: Ошибка при получении списка участников

  # Channel Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID комнаты
        name:
          type: string
          description: Название комнаты
        created_by:
          type: string
          format: uuid
          description: ID создателя комнаты
        invite_code:
          type: string
          description: Код приглашения в комнату
        is_active:
          type: boolean
          description: Активна ли комната
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RoomParticipant:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID записи участника
        room_id:
          type: string
          format: uuid
          description: ID комнаты
        user_id:
          type: string
          format: uuid
          description: ID пользователя
        is_admin:
          type: boolean
          description: Является ли администратором комнаты
        joined_at:
          type: string
          format: date-time
          description: Время присоединения к комнате
        left_at:
          type: string
          format: date-time
          description: Время выхода из комнаты
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...